requirements.txt
google-auth==0.4.0
google-auth-httplib2==0.0.2
google-cloud-core==0.22.1
google-cloud-translate==0.22.0
googleapis-common-protos==1.5.0
httplib2==0.9.2
protobuf==3.1.0.post1
pyasn1==0.1.9
pyasn1-modules==0.0.8
rsa==3.4.2
six==1.10.0
wheel==0.24.0
